+++ Changelog.md
@@ -10,6 +10,7 @@
 * Only show webling booking button if configured
 * Removed google books PHP library, replaced with Javascript solution
 * Replaced lity lightbox with fsightbox which has a gallery feature
+* Separated receipt icon in transactions view and added progress icon, fixed error handling
 
 ## 2.29.0